+++ src/components/SellerDashboard/SellerInfo/index.ts
@@ -1,13 +1,24 @@
 // src/components/SellerDashboard/SellerInfo/index.ts
 
-// 컴포넌트 export
+// ==================== 기본 컴포넌트 ====================
 export { BRAND_COLORS, PrimaryButton, SecondaryButton } from "./constants";
 export { ProgressCircle, FormField, PageHeader } from "./BasicComponents";
 export { ProfilePreviewCard, CompletionCard } from "./CardComponents";
+
+// ==================== 폼 관련 컴포넌트 ====================
 export { BasicInfoForm, FormActions } from "./FormComponents";
 export type { BasicInfoFormData } from "./FormComponents";
+
+// ==================== 분리된 폼 서브 컴포넌트 ====================
+export { default as AddressSearchModal } from "./AddressSearchModal";
+export { default as TagInput } from "./TagInput";
+export { default as ProfileImageUpload } from "./ProfileImageUpload";
+export { default as OperatingHours } from "./OperatingHours";
+export { default as AddressInputSection } from "./AddressInputSection";
+
+// ==================== 커스텀 훅 ====================
 export { useSellerInfo } from "./useSellerInfo";
 export type { SellerInfoData } from "./useSellerInfo";
 
-// 🎯 메인 페이지 - 올바른 경로로 수정!
+// ==================== 메인 페이지 ====================
 export { default } from "../../../pages/SellerDashboardPage/SellerInfoPage";
